Fix %zu device printf corruption in the temporary GPUDBG tracing
The H200 run's [GPUDBG] output was garbage: step numbers like
70369256030624 and every subsequent field shifted by one slot. CUDA's
device-side printf doesn't reliably support the 'z' length modifier for
size_t -- cast step_index to unsigned long long and use %llu instead,
which is the documented portable choice for CUDA device printf.

Also confirms the H200's real anomaly is unrelated to this printf bug:
the corruption is in what got displayed, not in the values the kernel
actually computed with (those come from separate Kokkos View reads).

+++ src/geodesic_integrate/integrate_cart_ks.hpp
@@ -164,8 +164,8 @@ struct Geodesic_cartesian_kerr_schild {
         // TEMPORARY: tracing why idx==0's state was observed bit-identical
         // across many steps on an H200 run -- remove once root-caused.
         if (idx == 0) {
-            printf("[GPUDBG] step=%zu enter idx=0 terminate=%d\n",
-                   step_index, static_cast<int>(photon_terminate(idx)));
+            printf("[GPUDBG] step=%llu enter idx=0 terminate=%d\n",
+                   static_cast<unsigned long long>(step_index), static_cast<int>(photon_terminate(idx)));
         }
         if (photon_terminate(idx)) {
             return;
@@ -182,8 +182,8 @@ struct Geodesic_cartesian_kerr_schild {
             photon_k3(idx)
         };
         if (idx == 0) {
-            printf("[GPUDBG] step=%zu idx=0 pre-state t=%.9f x=%.9f y=%.9f z=%.9f kt=%.9f kx=%.9f ky=%.9f kz=%.9f dlambda=%.9e integrator=%d\n",
-                   step_index, state[0], state[1], state[2], state[3],
+            printf("[GPUDBG] step=%llu idx=0 pre-state t=%.9f x=%.9f y=%.9f z=%.9f kt=%.9f kx=%.9f ky=%.9f kz=%.9f dlambda=%.9e integrator=%d\n",
+                   static_cast<unsigned long long>(step_index), state[0], state[1], state[2], state[3],
                    state[4], state[5], state[6], state[7],
                    photon_dlambda(idx), static_cast<int>(integrator));
         }
@@ -230,8 +230,8 @@ struct Geodesic_cartesian_kerr_schild {
         }
 
         if (idx == 0) {
-            printf("[GPUDBG] step=%zu idx=0 post-integrate state t=%.9f x=%.9f y=%.9f z=%.9f kt=%.9f kx=%.9f ky=%.9f kz=%.9f dlambda=%.9e\n",
-                   step_index, state[0], state[1], state[2], state[3],
+            printf("[GPUDBG] step=%llu idx=0 post-integrate state t=%.9f x=%.9f y=%.9f z=%.9f kt=%.9f kx=%.9f ky=%.9f kz=%.9f dlambda=%.9e\n",
+                   static_cast<unsigned long long>(step_index), state[0], state[1], state[2], state[3],
                    state[4], state[5], state[6], state[7], photon_dlambda(idx));
         }
